leaflet layer control sorta working
this needs to be cleaned and streamlined quite a bit and i have some
ideas, just tired of it for now.

Found that there are some issues with the ESRI legend, like the
"wetlands" heading appears twice.  I'm having the same thing happening.
Can probably program around it with some more work.

But potentially there may just be too many options for ESRI dynamic map
server legends to account for.