Updates to layers object and handlers
Sever updates to get the layers object fully functioning in the esri
template. Fixed radio button function. Updated TOC building to honor
visible setting in the layer object as well as other properties in the
case of a feature layer and visible layers for dynamic layers. Updated
text for group headings to reflect specifics about the examples layer
below.